agrego ruta, paginas y comp de turnos, modifico estilos en paginas y css, añado img de patrocinadores y agrego register comp
