+++ src/css/style.css
@@ -75,6 +75,22 @@ footer {
   -webkit-box-shadow: 0px 1px 20px 9px rgba(63, 63, 63, 0.82);
   -moz-box-shadow: 0px 1px 20px 9px rgba(65, 65, 65, 0.82);
 }
+@media (max-width: 550px) {
+  .turnForm {
+    display: flex;
+    flex-direction: column;
+    text-align: center;
+  }
+  .turnFormFecha {
+    width: 100%;
+  }
+}
+
+@media (min-width: 551px){
+  .turnFormFecha {
+    width: 50%;
+  }
+}
 @media (min-width: 992px) {
   .displayNone {
     display: none;
