+++ src/pages/HomePage.jsx
@@ -1,7 +1,5 @@
 import React from "react";
 import { Col, Container, Row } from "react-bootstrap";
-import Button from "react-bootstrap/Button";
-import Card from "react-bootstrap/Card";
 import Carousel from "react-bootstrap/Carousel";
 import CardComp from "../components/CardComp";
 
@@ -14,17 +12,17 @@ const HomePage = () => {
           <Col lg={6} md={12} sm={12}>
             <h2>Productos Destacados</h2>
             <hr />
-            <div className="row">
+            <Row className="justify-content-center">
               <CardComp />
               <CardComp />
               <CardComp />
-            </div>
+            </Row>
           </Col>
 
           <Col lg={6} md={12} sm={12}>
             <h2>Servicios</h2>
             <hr />
-            <Row>
+            <Row className="justify-content-center">
               <CardComp />
               <CardComp />
               <CardComp />
@@ -42,66 +40,47 @@ const HomePage = () => {
           </Col>
 
           <Col lg={6} md={12} sm={12}>
-            <h2 className="text-center mt-3">Publicidad</h2>
+            <h2 className="mt-3">Publicidad</h2>
             <hr />
+            <div className="d-flex justify-content-center">
+
             <Carousel data-bs-theme="dark">
               <Carousel.Item>
                 <img
-                  className="d-block w-50"
-                  src="Error404.png"
+                  className="d-block"
+                  src="/patrocinadores/purina.png"
                   alt="First slide"
                 />
-                <Carousel.Caption>
-                  <h5>First slide label</h5>
-                  <p>
-                    Nulla vitae elit libero, a pharetra augue mollis interdum.
-                  </p>
-                </Carousel.Caption>
               </Carousel.Item>
               <Carousel.Item>
                 <img
-                  className="d-block w-50"
-                  src="Error404.png"
+                  className="d-block"
+                  src="/patrocinadores/royalCanin.png"
                   alt="Second slide"
                 />
-                <Carousel.Caption>
-                  <h5>Second slide label</h5>
-                  <p>
-                    Lorem ipsum dolor sit amet, consectetur adipiscing elit.
-                  </p>
-                </Carousel.Caption>
               </Carousel.Item>
               <Carousel.Item>
                 <img
-                  className="d-block w-50"
-                  src="Error404.png"
+                  className="d-block"
+                  src="/patrocinadores/whiskas.png"
                   alt="Third slide"
                 />
-                <Carousel.Caption>
-                  <h5>Third slide label</h5>
-                  <p>
-                    Praesent commodo cursus magna, vel scelerisque nisl
-                    consectetur.
-                  </p>
-                </Carousel.Caption>
               </Carousel.Item>
             </Carousel>
+            </div>
           </Col>
-          <Col>
-            <h2 className="text-center mt-3">Comentarios</h2>
+          <Col lg={6} md={12} sm={12}>
+            <h2 className="mt-3">Comentarios</h2>
             <hr />
-            <Row>
-
-            <aside className="bg-info-subtle my-2  p-3 w-75 rounded-3 sombra" >
+            <div className="bg-info-subtle my-2  p-3 w-75 rounded-3 sombra">
               Comentario
-            </aside>
-            <aside className="bg-info-subtle  my-2 p-3 w-75 rounded-3 sombra">
+            </div>
+            <div className="bg-info-subtle  my-2 p-3 w-75 rounded-3 sombra">
               Comentario
-            </aside>
-            <aside className="bg-info-subtle  my-2 p-3 w-75 rounded-3 sombra">
+            </div>
+            <div className="bg-info-subtle  my-2 p-3 w-75 rounded-3 sombra">
               Comentario
-            </aside>
-            </Row>
+            </div>
           </Col>
         </Row>
       </Container>
